+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Use @Rob--W's Content-Disposition parser (#73)
   (source: Rob--W/open-in-browser)[https://github.com/Rob--W/open-in-browser/blob/master/extension/content-disposition.js]
+* Add localisation hooks to most things
 
 # 2.6.2